refactor: ombulCreate
