--- src/actions/ombul/Types.ts
@@ -1,7 +1,0 @@
-import type { CmsImage, Image, Ombul } from '@prisma/client'
-
-export type ExpandedOmbul = Ombul & {
-    coverImage: CmsImage & {
-        image: Image | null
-    }
-}
