+++ src/actions/ombul/create.ts
@@ -1,22 +1,18 @@
 'use server'
 import { createOmbulSchema } from './schema'
-import { createActionError, createPrismaActionError, createZodActionError } from '@/actions/error'
-import { readSpecialImageCollection } from '@/server/images/collections/read'
-import { createCmsImage } from '@/cms/images/create'
-import prisma from '@/prisma'
-import createFile from '@/server/store/createFile'
-import { createImage } from '@/server/images/create'
+import { createActionError, createZodActionError } from '@/actions/error'
 import { getUser } from '@/auth/user'
 import type { ActionReturn } from '@/actions/Types'
 import type { Ombul } from '@prisma/client'
 import type { CreateOmbulSchemaType } from './schema'
+import { createOmbul } from '@/server/ombul/create'
 
 /**
  * Create a new Ombul.
  * @param rawData includes a pdf file with the ombul issue optionaly year and issueNumber
  * @param CoverImageId is the id of the Image that will be used as the cover of the ombul
  */
-export async function createOmbul(rawdata: FormData | CreateOmbulSchemaType): Promise<ActionReturn<Ombul>> {
+export async function createOmbulAction(rawdata: FormData | CreateOmbulSchemaType): Promise<ActionReturn<Ombul>> {
     //Auth route
     const { status, authorized } = await getUser({
         requiredPermissions: ['OMBUL_CREATE']
@@ -29,75 +25,7 @@ export async function createOmbul(rawdata: FormData | CreateOmbulSchemaType): Pr
     if (!parse.success) {
         return createZodActionError(parse)
     }
-    const data = parse.data
+    const { ombulFile, ombulCoverImage, ...data } = parse.data
 
-    const year = data.year || new Date().getFullYear()
-
-    // Get the latest issue number if not provided
-    let latestIssueNumber = 1
-    if (!data.issueNumber) {
-        try {
-            const ombul = await prisma.ombul.findFirst({
-                where: {
-                    year
-                },
-                orderBy: {
-                    issueNumber: 'desc'
-                }
-            })
-            if (ombul) {
-                latestIssueNumber = ombul.issueNumber + 1
-            }
-        } catch (error) {
-            return createPrismaActionError(error)
-        }
-    }
-    const issueNumber = data.issueNumber || latestIssueNumber
-
-    const name = data.name
-    const description = data.description
-
-    //upload the file to the store volume
-    const ret = await createFile(data.ombulFile, 'ombul', ['pdf'])
-    if (!ret.success) return ret
-    const fsLocation = ret.data.fsLocation
-
-    // create coverimage
-    const ombulCoverCollectionRes = await readSpecialImageCollection('OMBULCOVERS')
-    if (!ombulCoverCollectionRes.success) return ombulCoverCollectionRes
-    const ombulCoverCollection = ombulCoverCollectionRes.data
-    const coverImageRes = await createImage(data.ombulCoverImage, {
-        name: fsLocation,
-        alt: `cover of ${name}`,
-        collectionId: ombulCoverCollection.id
-    })
-    if (!coverImageRes.success) return coverImageRes
-    const coverImage = coverImageRes.data
-
-    const cmsCoverImageRes = await createCmsImage(fsLocation, coverImage)
-    if (!cmsCoverImageRes.success) return cmsCoverImageRes
-    const cmsCoverImage = cmsCoverImageRes.data
-
-    try {
-        const ombul = await prisma.ombul.create({
-            data: {
-                description,
-                year,
-                issueNumber,
-                name,
-                coverImage: {
-                    connect: {
-                        id: cmsCoverImage.id
-                    }
-                },
-                fsLocation,
-            }
-        })
-        return {
-            success: true,
-            data: ombul
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    return await createOmbul(ombulFile, ombulCoverImage, data)
 }
