+++ src/app/ombul/page.tsx
@@ -5,7 +5,7 @@ import PageWrapper from '@/components/PageWrapper/PageWrapper'
 import AddHeaderItemPopUp from '@/components/AddHeaderItem/AddHeaderItemPopUp'
 import { readLatestOmbul, readOmbuls } from '@/actions/ombul/read'
 import { getUser } from '@/auth/user'
-import type { ExpandedOmbul } from '@/actions/ombul/Types'
+import type { ExpandedOmbul } from '@/server/ombul/Types'
 
 export default async function Ombuls() {
     const { user } = await getUser({
